requirements.txt
hata
python-dotenv
requests